requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
Deprecated==1.2.12
Django==3.2.4
django-bootstrap-v5==1.0.4
djangorestframework==3.12.4
idna==3.2
importlib-metadata==3.10.1
jwcrypto==1.0
Markdown==3.3.4
oauthlib==3.1.1
Pillow==8.2.0
psycopg2-binary==2.9.1
pycparser==2.20
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
wrapt==1.12.1
zipp==3.4.1
